// src/enums/Feature.ts
/**
 * Enum Feature provides standardised feature names.
 */
enum Feature {
  Assist = 'assist',
  Access = 'access',
  Alternatives = 'alternatives',
  SearchEnhancer = 'searchEnhancer',
  Redirected = 'redirected',
  TocAlerts = 'tocAlerts',
  Scite = 'scite',
  Modal = 'modal',
  LibrarySearch = 'librarySearch',
  LibraryServices = 'libraryServices',
  Campaign = 'campaign',
  KeywordEnhancements = 'keywordEnhancements',
  AlternativesPhysical = 'alternativesPhysical',
  SystemMessage = 'systemMessage',
}

export default Feature;